apps/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
freezegun
httpx
requests
passlib[bcrypt]
python-jose[cryptography]
pymysql
python-multipart
stripe
sentry-sdk
